requirements.txt
Django==1.6.7
django-hstore==1.3.5
psycopg2==2.5.4
iso8601==0.1.10
django-oauth2-provider==0.2.6.1
South==1.0
django-model-utils==2.0.1
djangorestframework==2.4.4
djangorestframework-gis==0.7
moment==0.2.2
django-simple-history==1.5.0
django-aggregate-if==0.4
django_braces==1.4.0
pillow==2.6.1
django_nose>=1.2
factory-boy==2.4.1
pytz==2014.7
django-youtube==0.2
gdata==2.0.18
easy-thumbnails==2.2
requests==2.4.3
